+++ compiler/extensions/cpp/runtime/test/zserio/ArrayTest.cpp
@@ -700,49 +700,61 @@ private:
 TEST_F(ArrayTest, intField4Array)
 {
     const size_t numBits = 4;
-    std::vector<int8_t> rawArray = {-(1 << (numBits - 1)), 7, (1 << (numBits - 1)) - 1};
+    std::vector<int8_t> rawArray = {
+            -static_cast<int8_t>(1U << (numBits - 1)),
+            7,
+            static_cast<int8_t>(1U << (numBits - 1)) - 1};
     testArray(rawArray, BitFieldArrayTraits<int8_t>(numBits), numBits);
 }
 
 TEST_F(ArrayTest, intField12Array)
 {
     const size_t numBits = 12;
-    std::vector<int16_t> rawArray = {-(1 << (numBits - 1)), 7, (1 << (numBits - 1)) - 1};
+    std::vector<int16_t> rawArray = {
+            -static_cast<int16_t>(1U << (numBits - 1)),
+            7,
+            static_cast<int16_t>(1U << (numBits - 1)) - 1};
     testArray(rawArray, BitFieldArrayTraits<int16_t>(numBits), numBits);
 }
 
 TEST_F(ArrayTest, intField20Array)
 {
     const size_t numBits = 20;
-    std::vector<int32_t> rawArray = {-(1 << (numBits - 1)), 7, (1 << (numBits - 1)) - 1};
+    std::vector<int32_t> rawArray = {
+            -static_cast<int32_t>(1U << (numBits - 1)),
+            7,
+            static_cast<int32_t>(1U << (numBits - 1)) - 1};
     testArray(rawArray, BitFieldArrayTraits<int32_t>(numBits), numBits);
 }
 
 TEST_F(ArrayTest, intField36Array)
 {
     const size_t numBits = 36;
-    std::vector<int64_t> rawArray = {-(INT64_C(1) << (numBits - 1)), 7, (INT64_C(1) << (numBits - 1)) - 1};
+    std::vector<int64_t> rawArray = {
+            -static_cast<int64_t>(UINT64_C(1) << (numBits - 1)),
+            7,
+            static_cast<int64_t>(UINT64_C(1) << (numBits - 1)) - 1};
     testArray(rawArray, BitFieldArrayTraits<int64_t>(numBits), numBits);
 }
 
 TEST_F(ArrayTest, bitField4Array)
 {
     const size_t numBits = 4;
-    std::vector<uint8_t> rawArray = {0, 7, (1 << numBits) - 1};
+    std::vector<uint8_t> rawArray = {0, 7, (1U << numBits) - 1};
     testArray(rawArray, BitFieldArrayTraits<uint8_t>(numBits), numBits);
 }
 
 TEST_F(ArrayTest, bitField12Array)
 {
     const size_t numBits = 12;
-    std::vector<uint16_t> rawArray = {0, 7, (1 << numBits) - 1};
+    std::vector<uint16_t> rawArray = {0, 7, (1U << numBits) - 1};
     testArray(rawArray, BitFieldArrayTraits<uint16_t>(numBits), numBits);
 }
 
 TEST_F(ArrayTest, bitField20Array)
 {
     const size_t numBits = 20;
-    std::vector<uint32_t> rawArray = {0, 7, (1 << numBits) - 1};
+    std::vector<uint32_t> rawArray = {0, 7, (1U << numBits) - 1};
     testArray(rawArray, BitFieldArrayTraits<uint32_t>(numBits), numBits);
 }
 
@@ -756,7 +768,10 @@ TEST_F(ArrayTest, bitField36Array)
 TEST_F(ArrayTest, dynamicIntField4Array)
 {
     const size_t numBits = 4;
-    std::vector<int8_t> rawArray = {-(1 << (numBits - 1)), 7, (1 << (numBits - 1)) - 1};
+    std::vector<int8_t> rawArray = {
+            -static_cast<int8_t>(1U << (numBits - 1)),
+            7,
+            static_cast<int8_t>(1U << (numBits - 1)) - 1};
     const ElementBitSize elementBitSize(numBits);
     testArray(rawArray, DynamicBitFieldArrayTraits<int8_t, ElementBitSize>(elementBitSize), numBits);
 }
@@ -764,7 +779,10 @@ TEST_F(ArrayTest, dynamicIntField4Array)
 TEST_F(ArrayTest, dynamicIntField12Array)
 {
     const size_t numBits = 12;
-    std::vector<int16_t> rawArray = {-(1 << (numBits - 1)), 7, (1 << (numBits - 1)) - 1};
+    std::vector<int16_t> rawArray = {
+            -static_cast<int16_t>(1U << (numBits - 1)),
+            7,
+            static_cast<int16_t>(1U << (numBits - 1)) - 1};
     const ElementBitSize elementBitSize(numBits);
     testArray(rawArray, DynamicBitFieldArrayTraits<int16_t, ElementBitSize>(elementBitSize), numBits);
 }
@@ -772,7 +790,10 @@ TEST_F(ArrayTest, dynamicIntField12Array)
 TEST_F(ArrayTest, dynamicIntField20Array)
 {
     const size_t numBits = 20;
-    std::vector<int32_t> rawArray = {-(1 << (numBits - 1)), 7, (1 << (numBits - 1)) - 1};
+    std::vector<int32_t> rawArray = {
+            -static_cast<int32_t>(1U << (numBits - 1)),
+            7,
+            static_cast<int32_t>(1U << (numBits - 1)) - 1};
     const ElementBitSize elementBitSize(numBits);
     testArray(rawArray, DynamicBitFieldArrayTraits<int32_t, ElementBitSize>(elementBitSize), numBits);
 }
@@ -780,7 +801,10 @@ TEST_F(ArrayTest, dynamicIntField20Array)
 TEST_F(ArrayTest, dynamicIntField36Array)
 {
     const size_t numBits = 36;
-    std::vector<int64_t> rawArray = {-(INT64_C(1) << (numBits - 1)), 7, (INT64_C(1) << (numBits - 1)) - 1};
+    std::vector<int64_t> rawArray = {
+            -static_cast<int64_t>(UINT64_C(1) << (numBits - 1)),
+            7,
+            static_cast<int64_t>(UINT64_C(1) << (numBits - 1)) - 1};
     const ElementBitSize elementBitSize(numBits);
     testArray(rawArray, DynamicBitFieldArrayTraits<int64_t, ElementBitSize>(elementBitSize), numBits);
 }
@@ -788,7 +812,7 @@ TEST_F(ArrayTest, dynamicIntField36Array)
 TEST_F(ArrayTest, dynamicBitField4Array)
 {
     const size_t numBits = 4;
-    std::vector<uint8_t> rawArray = {0, 7, (1 << numBits) - 1};
+    std::vector<uint8_t> rawArray = {0, 7, (1U << numBits) - 1};
     const ElementBitSize elementBitSize(numBits);
     testArray(rawArray, DynamicBitFieldArrayTraits<uint8_t, ElementBitSize>(elementBitSize), numBits);
 }
@@ -796,7 +820,7 @@ TEST_F(ArrayTest, dynamicBitField4Array)
 TEST_F(ArrayTest, dynamicBitField12Array)
 {
     const size_t numBits = 12;
-    std::vector<uint16_t> rawArray = {0, 7, (1 << numBits) - 1};
+    std::vector<uint16_t> rawArray = {0, 7, (1U << numBits) - 1};
     const ElementBitSize elementBitSize(numBits);
     testArray(rawArray, DynamicBitFieldArrayTraits<uint16_t, ElementBitSize>(elementBitSize), numBits);
 }
@@ -804,7 +828,7 @@ TEST_F(ArrayTest, dynamicBitField12Array)
 TEST_F(ArrayTest, dynamicBitField20Array)
 {
     const size_t numBits = 20;
-    std::vector<uint32_t> rawArray = {0, 7, (1 << numBits) - 1};
+    std::vector<uint32_t> rawArray = {0, 7, (1U << numBits) - 1};
     const ElementBitSize elementBitSize(numBits);
     testArray(rawArray, DynamicBitFieldArrayTraits<uint32_t, ElementBitSize>(elementBitSize), numBits);
 }
@@ -867,14 +891,18 @@ TEST_F(ArrayTest, stdUInt64Array)
 
 TEST_F(ArrayTest, varInt16Array)
 {
-    std::vector<int16_t> rawArray = {1 << 5, 1 << (5 + 8)};
+    std::vector<int16_t> rawArray = {static_cast<int16_t>(1U << 5U), static_cast<int16_t>(1U << (5U + 8))};
     const size_t bitSize = 8 * (1 + 2);
     testArray(rawArray, VarIntNNArrayTraits<int16_t>(), bitSize, bitSize);
 }
 
 TEST_F(ArrayTest, varInt32Array)
 {
-    std::vector<int32_t> rawArray = {1 << 5, 1 << (5 + 7), 1 << (5 + 7 + 7), 1 << (5 + 7 + 7 + 8)};
+    std::vector<int32_t> rawArray = {
+            static_cast<int32_t>(1U << 5U),
+            static_cast<int32_t>(1U << (5U + 7)),
+            static_cast<int32_t>(1U << (5U + 7 + 7)),
+            static_cast<int32_t>(1U << (5U + 7 + 7 + 8))};
     const size_t bitSize = 8 * (1 + 2 + 3 + 4);
     testArray(rawArray, VarIntNNArrayTraits<int32_t>(), bitSize, bitSize);
 }
@@ -882,28 +910,28 @@ TEST_F(ArrayTest, varInt32Array)
 TEST_F(ArrayTest, varInt64Array)
 {
     std::vector<int64_t> rawArray = {
-            INT64_C(1) << 5,
-            INT64_C(1) << (5 + 7),
-            INT64_C(1) << (5 + 7 + 7),
-            INT64_C(1) << (5 + 7 + 7 + 7),
-            INT64_C(1) << (5 + 7 + 7 + 7 + 7),
-            INT64_C(1) << (5 + 7 + 7 + 7 + 7 + 7),
-            INT64_C(1) << (5 + 7 + 7 + 7 + 7 + 7 + 7),
-            INT64_C(1) << (5 + 7 + 7 + 7 + 7 + 7 + 7 + 8)};
+            static_cast<int64_t>(UINT64_C(1) << 5U),
+            static_cast<int64_t>(UINT64_C(1) << (5U + 7)),
+            static_cast<int64_t>(UINT64_C(1) << (5U + 7 + 7)),
+            static_cast<int64_t>(UINT64_C(1) << (5U + 7 + 7 + 7)),
+            static_cast<int64_t>(UINT64_C(1) << (5U + 7 + 7 + 7 + 7)),
+            static_cast<int64_t>(UINT64_C(1) << (5U + 7 + 7 + 7 + 7 + 7)),
+            static_cast<int64_t>(UINT64_C(1) << (5U + 7 + 7 + 7 + 7 + 7 + 7)),
+            static_cast<int64_t>(UINT64_C(1) << (5U + 7 + 7 + 7 + 7 + 7 + 7 + 8))};
     const size_t bitSize = 8 * (1 + 2 + 3 + 4 + 5 + 6 + 7 + 8);
     testArray(rawArray, VarIntNNArrayTraits<int64_t>(), bitSize, bitSize);
 }
 
 TEST_F(ArrayTest, varUInt16Array)
 {
-    std::vector<uint16_t> rawArray = {1 << 6, 1 << (6 + 8)};
+    std::vector<uint16_t> rawArray = {1U << 6U, 1U << (6U + 8)};
     const size_t bitSize = 8 * (1 + 2);
     testArray(rawArray, VarIntNNArrayTraits<uint16_t>(), bitSize, bitSize);
 }
 
 TEST_F(ArrayTest, varUInt32Array)
 {
-    std::vector<uint32_t> rawArray = {1 << 6, 1 << (6 + 7), 1 << (6 + 7 + 7), 1 << (6 + 7 + 7 + 8)};
+    std::vector<uint32_t> rawArray = {1U << 6U, 1U << (6U + 7), 1U << (6U + 7 + 7), 1U << (6U + 7 + 7 + 8)};
     const size_t bitSize = 8 * (1 + 2 + 3 + 4);
     testArray(rawArray, VarIntNNArrayTraits<uint32_t>(), bitSize, bitSize);
 }
@@ -911,14 +939,14 @@ TEST_F(ArrayTest, varUInt32Array)
 TEST_F(ArrayTest, varUInt64Array)
 {
     std::vector<uint64_t> rawArray = {
-            UINT64_C(1) << 6,
-            UINT64_C(1) << (6 + 7),
-            UINT64_C(1) << (6 + 7 + 7),
-            UINT64_C(1) << (6 + 7 + 7 + 7),
-            UINT64_C(1) << (6 + 7 + 7 + 7 + 7),
-            UINT64_C(1) << (6 + 7 + 7 + 7 + 7 + 7),
-            UINT64_C(1) << (6 + 7 + 7 + 7 + 7 + 7 + 7),
-            UINT64_C(1) << (6 + 7 + 7 + 7 + 7 + 7 + 7 + 8)};
+            UINT64_C(1) << 6U,
+            UINT64_C(1) << (6U + 7),
+            UINT64_C(1) << (6U + 7 + 7),
+            UINT64_C(1) << (6U + 7 + 7 + 7),
+            UINT64_C(1) << (6U + 7 + 7 + 7 + 7),
+            UINT64_C(1) << (6U + 7 + 7 + 7 + 7 + 7),
+            UINT64_C(1) << (6U + 7 + 7 + 7 + 7 + 7 + 7),
+            UINT64_C(1) << (6U + 7 + 7 + 7 + 7 + 7 + 7 + 8)};
     const size_t bitSize = 8 * (1 + 2 + 3 + 4 + 5 + 6 + 7 + 8);
     testArray(rawArray, VarIntNNArrayTraits<uint64_t>(), bitSize, bitSize);
 }
@@ -930,29 +958,29 @@ TEST_F(ArrayTest, varIntArray)
     rawArray.push_back(0);
     rawArray.push_back(-1);
     rawArray.push_back(1);
-    rawArray.push_back(-(INT64_C(1) << 6) + 1);
-    rawArray.push_back((INT64_C(1) << 6) - 1);
+    rawArray.push_back(-static_cast<int64_t>(UINT64_C(1) << 6U) + 1);
+    rawArray.push_back(static_cast<int64_t>(UINT64_C(1) << 6U) - 1);
     // 2 bytes
-    rawArray.push_back(-(INT64_C(1) << 13) + 1);
-    rawArray.push_back((INT64_C(1) << 13) - 1);
+    rawArray.push_back(-static_cast<int64_t>(UINT64_C(1) << 13U) + 1);
+    rawArray.push_back(static_cast<int64_t>(UINT64_C(1) << 13U) - 1);
     // 3 bytes
-    rawArray.push_back(-(INT64_C(1) << 20) + 1);
-    rawArray.push_back((INT64_C(1) << 20) - 1);
+    rawArray.push_back(-static_cast<int64_t>(UINT64_C(1) << 20U) + 1);
+    rawArray.push_back(static_cast<int64_t>(UINT64_C(1) << 20U) - 1);
     // 4 bytes
-    rawArray.push_back(-(INT64_C(1) << 27) + 1);
-    rawArray.push_back((INT64_C(1) << 27) - 1);
+    rawArray.push_back(-static_cast<int64_t>(UINT64_C(1) << 27U) + 1);
+    rawArray.push_back(static_cast<int64_t>(UINT64_C(1) << 27U) - 1);
     // 5 bytes
-    rawArray.push_back(-(INT64_C(1) << 34) + 1);
-    rawArray.push_back((INT64_C(1) << 34) - 1);
+    rawArray.push_back(-static_cast<int64_t>(UINT64_C(1) << 34U) + 1);
+    rawArray.push_back(static_cast<int64_t>(UINT64_C(1) << 34U) - 1);
     // 6 bytes
-    rawArray.push_back(-(INT64_C(1) << 41) + 1);
-    rawArray.push_back((INT64_C(1) << 41) - 1);
+    rawArray.push_back(-static_cast<int64_t>(UINT64_C(1) << 41U) + 1);
+    rawArray.push_back(static_cast<int64_t>(UINT64_C(1) << 41U) - 1);
     // 7 bytes
-    rawArray.push_back(-(INT64_C(1) << 48) + 1);
-    rawArray.push_back((INT64_C(1) << 48) - 1);
+    rawArray.push_back(-static_cast<int64_t>(UINT64_C(1) << 48U) + 1);
+    rawArray.push_back(static_cast<int64_t>(UINT64_C(1) << 48U) - 1);
     // 8 bytes
-    rawArray.push_back(-(INT64_C(1) << 55) + 1);
-    rawArray.push_back((INT64_C(1) << 55) - 1);
+    rawArray.push_back(-static_cast<int64_t>(UINT64_C(1) << 55U) + 1);
+    rawArray.push_back(static_cast<int64_t>(UINT64_C(1) << 55U) - 1);
     // 9 bytes
     rawArray.push_back(INT64_MIN + 1);
     rawArray.push_back(INT64_MAX);
@@ -968,21 +996,21 @@ TEST_F(ArrayTest, varUIntArray)
     // 1 byte
     rawArray.push_back(0);
     rawArray.push_back(1);
-    rawArray.push_back((UINT64_C(1) << 7) - 1);
+    rawArray.push_back((UINT64_C(1) << 7U) - 1);
     // 2 bytes
-    rawArray.push_back((UINT64_C(1) << 14) - 1);
+    rawArray.push_back((UINT64_C(1) << 14U) - 1);
     // 3 bytes
-    rawArray.push_back((UINT64_C(1) << 21) - 1);
+    rawArray.push_back((UINT64_C(1) << 21U) - 1);
     // 4 bytes
-    rawArray.push_back((UINT64_C(1) << 28) - 1);
+    rawArray.push_back((UINT64_C(1) << 28U) - 1);
     // 5 bytes
-    rawArray.push_back((UINT64_C(1) << 35) - 1);
+    rawArray.push_back((UINT64_C(1) << 35U) - 1);
     // 6 bytes
-    rawArray.push_back((UINT64_C(1) << 42) - 1);
+    rawArray.push_back((UINT64_C(1) << 42U) - 1);
     // 7 bytes
-    rawArray.push_back((UINT64_C(1) << 49) - 1);
+    rawArray.push_back((UINT64_C(1) << 49U) - 1);
     // 8 bytes
-    rawArray.push_back((UINT64_C(1) << 56) - 1);
+    rawArray.push_back((UINT64_C(1) << 56U) - 1);
     // 9 bytes
     rawArray.push_back(UINT64_MAX);
     const size_t bitSize = 8 * (2 + (1 + 2 + 3 + 4 + 5 + 6 + 7 + 8 + 9));
@@ -992,11 +1020,11 @@ TEST_F(ArrayTest, varUIntArray)
 TEST_F(ArrayTest, varSizeArray)
 {
     std::vector<uint32_t> rawArray = {
-            UINT32_C(1) << 6,
-            UINT32_C(1) << (6 + 7),
-            UINT32_C(1) << (6 + 7 + 7),
-            UINT32_C(1) << (6 + 7 + 7 + 7),
-            UINT32_C(1) << (1 + 7 + 7 + 7 + 8)};
+            UINT32_C(1) << 6U,
+            UINT32_C(1) << (6U + 7),
+            UINT32_C(1) << (6U + 7 + 7),
+            UINT32_C(1) << (6U + 7 + 7 + 7),
+            UINT32_C(1) << (1U + 7 + 7 + 7 + 8)};
     const size_t bitSize = 8 * (1 + 2 + 3 + 4 + 5);
     testArray(rawArray, VarSizeArrayTraits(), bitSize, bitSize);
 }
@@ -1192,14 +1220,14 @@ TEST_F(ArrayTest, intField16PackedArray)
 TEST_F(ArrayTest, varUInt64PackedArray)
 {
     std::vector<uint64_t> rawArray = {
-            UINT64_C(1) << 6,
-            UINT64_C(1) << (6 + 7),
-            UINT64_C(1) << (6 + 7 + 7),
-            UINT64_C(1) << (6 + 7 + 7 + 7),
-            UINT64_C(1) << (6 + 7 + 7 + 7 + 7),
-            UINT64_C(1) << (6 + 7 + 7 + 7 + 7 + 7),
-            UINT64_C(1) << (6 + 7 + 7 + 7 + 7 + 7 + 7),
-            UINT64_C(1) << (6 + 7 + 7 + 7 + 7 + 7 + 7 + 8)};
+            UINT64_C(1) << 6U,
+            UINT64_C(1) << (6U + 7),
+            UINT64_C(1) << (6U + 7 + 7),
+            UINT64_C(1) << (6U + 7 + 7 + 7),
+            UINT64_C(1) << (6U + 7 + 7 + 7 + 7),
+            UINT64_C(1) << (6U + 7 + 7 + 7 + 7 + 7),
+            UINT64_C(1) << (6U + 7 + 7 + 7 + 7 + 7 + 7),
+            UINT64_C(1) << (6U + 7 + 7 + 7 + 7 + 7 + 7 + 8)};
     testPackedArray(rawArray, VarIntNNArrayTraits<uint64_t>());
 
     std::vector<uint64_t> unpackedRawArray = {UINT64_C(5000000), 0, 0, 0, 0, 0, 0};
